+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed 
+- throttledEvent now transfers the instance of the event to the callback.
 
 ## [1.0.2] - 2018-08-29
 ### Added